+++ lending/Codat/Lending/Models/Components/ReportType.cs
@@ -17,6 +17,8 @@ namespace Codat.Lending.Models.Components
     {
         [JsonProperty("categorizedBankStatement")]
         CategorizedBankStatement,
+        [JsonProperty("creditModel")]
+        CreditModel,
     }
 
     public static class ReportTypeExtension