+++ lending/Codat/Lending/Models/Webhooks/ClientRateLimitReachedResponse1.cs
@@ -10,8 +10,8 @@
 namespace Codat.Lending.Models.Webhooks
 {
     using Codat.Lending.Utils;
-    using System.Net.Http;
     using System;
+    using System.Net.Http;
     
     public class ClientRateLimitReachedResponse1
     {